pyproject.toml
[project]
name = "llm-app"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "notebook>=7.4.7",
    "pillow>=11.3.0",
    "torch>=2.8.0",
    "torchtext>=0.18.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.1",
]
